test: racing game end condition
