+++ src/main/java/racingcar/model/RacingGame.java
@@ -47,12 +47,12 @@ public class RacingGame {
         trialCnt.increase();
     }
 
-    public Cars getCars() {
-        return cars;
-    }
-
     public Result findWinner() {
         int maxPosition = cars.findMaxPosition();
         return new Result(cars.findWinnerName(maxPosition), maxPosition);
+    }
+
+    public Cars getRacingCars() {
+        return cars;
     }
 }
