+++ src/test/java/racingcar/RacingGameTest.java
@@ -2,10 +2,10 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.controller.GameController;
 import racingcar.model.Car;
 import racingcar.model.Cars;
 import racingcar.model.Name;
+import racingcar.model.RacingGame;
 import racingcar.util.NumberGenerator;
 import racingcar.util.RandomNumberGenerator;
 
@@ -17,15 +17,14 @@ import java.util.stream.Collectors;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class RacingGameTest {
-    private static final int TESTCASE_NUM = 10000;
 
     @Test
-    @DisplayName("랜덤 숫자 생성")
-    void generateRandomNumber() {
-        final NumberGenerator randomNumber = new RandomNumberGenerator();
-        for (int i = 0; i < TESTCASE_NUM; i++) {
-            assertThat(randomNumber.generate()).isBetween(0,9);
-        }
+    @DisplayName("경주 게임 정상 종료 조건")
+    void checkEndRaceCondition() {
+        RacingGame racingGame = new RacingGame("aaa", "1");
+        racingGame.race();
+
+        assertThat(racingGame.isEnd()).isTrue();
     }
 
     @Test
